+++ swing/src/net/sf/openrocket/gui/main/flightconfigpanel/FlightConfigurablePanel.java
@@ -3,6 +3,8 @@ package net.sf.openrocket.gui.main.flightconfigpanel;
 import java.awt.Color;
 import java.awt.Component;
 import java.awt.Font;
+import java.awt.event.FocusEvent;
+import java.awt.event.FocusListener;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
 import java.util.Arrays;
@@ -180,6 +182,25 @@ public abstract class FlightConfigurablePanel<T extends FlightConfigurableCompon
 				}
 			}
 		});
+
+		table.getColumnModel().getSelectionModel().addListSelectionListener(new ListSelectionListener() {
+			@Override
+			public void valueChanged(ListSelectionEvent e) {
+				updateComponentSelection(e);
+			}
+		});
+
+		table.addFocusListener(new FocusListener() {
+			@Override
+			public void focusGained(FocusEvent e) {
+				updateComponentSelection(new ListSelectionEvent(this, 0, 0, false));
+			}
+
+			@Override
+			public void focusLost(FocusEvent e) {
+
+			}
+		});
 	}
 
 	/**
@@ -192,6 +213,11 @@ public abstract class FlightConfigurablePanel<T extends FlightConfigurableCompon
 	public void clearSelection() {
 		table.clearSelection();
 	}
+
+	/**
+	 * Update the selection in the rocket design view, based on the currently selected motor, recovery device, or stage.
+	 */
+	protected abstract void updateComponentSelection(ListSelectionEvent e);
 
 	protected T getSelectedComponent() {
